wip implementing the layout loader
