+++ client/ayon_unreal/api/plugin.py
@@ -9,6 +9,7 @@ from abc import (
 )
 
 import unreal
+import ayon_api
 
 from .pipeline import (
     create_publish_instance,
@@ -25,7 +26,12 @@ from ayon_core.pipeline import (
     Creator,
     LoaderPlugin,
     CreatorError,
-    CreatedInstance
+    CreatedInstance,
+    discover_loader_plugins,
+    loaders_from_representation,
+    load_container,
+    AYON_CONTAINER_ID,
+    get_current_project_name
 )
 
 
@@ -281,3 +287,180 @@ class UnrealActorCreator(UnrealBaseCreator):
 class Loader(LoaderPlugin, ABC):
     """This serves as skeleton for future Ayon specific functionality"""
     pass
+
+
+class LayoutLoader(Loader):
+    """Load Layout from a JSON file"""
+
+    product_types = {"layout"}
+    representations = {"json"}
+
+    label = "Load Layout"
+    icon = "code-fork"
+    color = "orange"
+
+    @staticmethod
+    def _get_fbx_loader(loaders, family):
+        name = ""
+        if family in ['rig', 'skeletalMesh']:
+            name = "SkeletalMeshFBXLoader"
+        elif family in ['model', 'staticMesh']:
+            name = "StaticMeshFBXLoader"
+        elif family == 'camera':
+            name = "CameraLoader"
+
+        if name == "":
+
+            return None
+
+        for loader in loaders:
+            if loader.__name__ == name:
+                return loader
+
+        return None
+
+    @staticmethod
+    def _get_abc_loader(loaders, family):
+        name = ""
+        if family in ['rig', 'skeletalMesh']:
+            name = "SkeletalMeshAlembicLoader"
+        elif family in ['model', 'staticMesh']:
+            name = "StaticMeshAlembicLoader"
+
+        if name == "":
+            return None
+
+        for loader in loaders:
+            if loader.__name__ == name:
+                return loader
+
+        return None
+
+    def _transform_from_basis(self, transform, basis):
+        """Transform a transform from a basis to a new basis."""
+        # Get the basis matrix
+        basis_matrix = unreal.Matrix(
+            basis[0],
+            basis[1],
+            basis[2],
+            basis[3]
+        )
+        transform_matrix = unreal.Matrix(
+            transform[0],
+            transform[1],
+            transform[2],
+            transform[3]
+        )
+
+        new_transform = (
+            basis_matrix.get_inverse() * transform_matrix * basis_matrix)
+
+        return new_transform.transform()
+
+    def _get_repre_entities_by_version_id(self, project_name, data, repre_extension, force_loaded=False):
+        version_ids = {
+            element.get("version")
+            for element in data
+            if element.get("representation")
+        }
+        version_ids.discard(None)
+        output = collections.defaultdict(list)
+        if not version_ids:
+            return output
+        # Extract extensions from data with backward compatibility for "ma"
+        extensions = {
+            element["extension"]
+            for element in data
+            if element.get("representation")
+        }
+
+        # Update extensions based on the force_loaded flag
+        updated_extensions = set()
+
+        for ext in extensions:
+            if not force_loaded or repre_extension == "json":
+                if ext == "ma":
+                    updated_extensions.update({"fbx", "abc"})
+                else:
+                    updated_extensions.add(ext)
+            else:
+                updated_extensions.update({repre_extension})
+
+        repre_entities = ayon_api.get_representations(
+            project_name,
+            representation_names=updated_extensions,
+            version_ids=version_ids,
+            fields={"id", "versionId", "name"}
+        )
+        for repre_entity in repre_entities:
+            version_id = repre_entity["versionId"]
+            output[version_id].append(repre_entity)
+        return output
+
+    def imprint(
+        self,
+        context,
+        folder_path,
+        folder_name,
+        loaded_assets,
+        asset_dir,
+        asset_name,
+        container_name,
+        hierarchy_dir=None
+    ):
+        data = {
+            "schema": "ayon:container-2.0",
+            "id": AYON_CONTAINER_ID,
+            "asset": folder_name,
+            "folder_path": folder_path,
+            "namespace": asset_dir,
+            "container_name": container_name,
+            "asset_name": asset_name,
+            "loader": str(self.__class__.__name__),
+            "representation": context["representation"]["id"],
+            "parent": context["representation"]["versionId"],
+            "family": context["product"]["productType"],
+            "loaded_assets": loaded_assets,
+        }
+        if hierarchy_dir is not None:
+            data["master_directory"] = hierarchy_dir
+        imprint(
+            "{}/{}".format(asset_dir, container_name), data)
+
+    def _load_assets(self, instance_name, repre_id, product_type, repr_format):
+        all_loaders = discover_loader_plugins()
+        loaders = loaders_from_representation(
+            all_loaders, repre_id)
+
+        loader = None
+
+        if repr_format == 'fbx':
+            loader = self._get_fbx_loader(loaders, product_type)
+        elif repr_format == 'abc':
+            loader = self._get_abc_loader(loaders, product_type)
+
+        if not loader:
+            if repr_format == "ma":
+                msg = (
+                    f"No valid {product_type} loader found for {repre_id} ({repr_format}), "
+                    f"consider using {product_type} loader (fbx/abc) instead."
+                )
+                self.log.warning(msg)
+            else:
+                self.log.error(
+                    f"No valid loader found for {repre_id} "
+                    f"({repr_format}) "
+                    f"{product_type}")
+            return
+
+        options = {
+            # "asset_dir": asset_dir
+        }
+
+        assets = load_container(
+            loader,
+            repre_id,
+            namespace=instance_name,
+            options=options
+        )
+        return assets
